Split out fastdotcom into a component and a sensor platform
