+++ homeassistant/components/fastdotcom/services.yaml
@@ -1,0 +1,2 @@
+speedtest:
+  description: Immediately take a speedest with Fast.com
